Move SPA and tomo workflow endpoints
